+++ neural_compressor/adaptor/torch_utils/autoround/autoround.py
@@ -1,6 +1,3 @@
-#
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) 2023 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
 try:
     from neural_compressor.utils.utility import LazyImport
@@ -28,51 +24,16 @@ except:  # pragma: no cover
     import torch
 
     logger = logging.getLogger()
+
 import copy
 from collections import UserDict
 from functools import partial
-from typing import Union
 
 from torch.amp import autocast
-from torch.autograd import Function
 from torch.functional import F
 
 
-def get_module(model, key):
-    """Get module from model by key name.
-
-    Args:
-        model (torch.nn.Module): original model
-        key (str): module name to be replaced
-    """
-    module = model
-    name_list = key.split(".")
-    for name in name_list:
-        if hasattr(module, name):
-            module = getattr(module, name)
-            module = module
-    return module
-
-
-def set_module(model, key, new_module):
-    """Set new module into model by key name.
-
-    Args:
-        model (torch.nn.Module): original model
-        key (str): module name to be replaced
-        new_module (torch.nn.Module): new module to be inserted
-    """
-    module = model
-    name_list = key.split(".")
-    for name in name_list[:-1]:
-        if hasattr(module, name):
-            module = getattr(module, name)
-        else:
-            module = module
-    setattr(module, name_list[-1], new_module)
-
-
-def quant_weight_asym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sigmoid=False):
+def quant_weight_asym(weight, num_bits=4, v=0, min_scale=0, max_scale=0):
     """Quantizes and dequantizes weight asymmetrically.
 
     Args:
@@ -81,23 +42,17 @@ def quant_weight_asym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sig
         v: Rounding value perturbation
         min_scale: Minimum scale coefficient for weight
         max_scale: Maximum scale coefficient for weight
-        use_sigmoid: Boolean indicating whether to use sigmoid; found useful in some scenarios for adam
 
     Returns:
         Quantized and dequantized weight, scale, zero-point
     """
-
     maxq = torch.tensor(2**num_bits - 1)
-    zeros = torch.zeros(weight.shape[0], device=weight.device, dtype=weight.dtype)
+    zeros = torch.zeros(weight.shape[0], device=weight.device)
     if isinstance(min_scale, torch.Tensor):
         wmin_tmp = torch.minimum(weight.min(1)[0], zeros)
         wmax_tmp = torch.maximum(weight.max(1)[0], zeros)
-        if use_sigmoid:
-            wmin_tmp *= torch.sigmoid(8 * (min_scale + 0.5))
-            wmax_tmp *= torch.sigmoid(8 * (max_scale + 0.5))
-        else:
-            wmin_tmp *= torch.sigmoid(1.0 + min_scale)
-            wmax_tmp *= torch.sigmoid(1.0 + max_scale)
+        wmin_tmp *= min_scale + 1.0
+        wmax_tmp *= max_scale + 1.0
         wmax = torch.maximum(wmax_tmp, wmin_tmp)
         wmin = torch.minimum(wmax_tmp, wmin_tmp)
     else:
@@ -115,7 +70,7 @@ def quant_weight_asym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sig
     return scale * (q - zp), scale, zp
 
 
-def quant_weight_sym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sigmoid=False):
+def quant_weight_sym(weight, num_bits=4, v=0, min_scale=0, max_scale=0):
     """Quantizes and dequantizes weight symmetrically.
 
     Args:
@@ -124,7 +79,6 @@ def quant_weight_sym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sigm
         v: Rounding value perturbation
         min_scale: Minimum scale coefficient for weight
         max_scale: Maximum scale coefficient for weight
-        use_sigmoid: Boolean indicating whether to use sigmoid; found useful in some scenarios for adam
 
     Returns:
         Quantized and dequantized weight, scale, zero-point
@@ -136,10 +90,7 @@ def quant_weight_sym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sigm
         minq = torch.tensor(2 ** (num_bits - 1) - 1)
 
     wmax = torch.abs(weight).max(1)[0]
-    if use_sigmoid:
-        wmax *= torch.sigmoid(8 * (min_scale + 0.5))
-    else:
-        wmax *= 1.0 + max_scale
+    wmax *= 1 + max_scale
     tmp = wmax == 0
     wmax[tmp] = +1
     scale = wmax / ((maxq - minq) / 2)
@@ -148,7 +99,7 @@ def quant_weight_sym(weight, num_bits=4, v=0, min_scale=0, max_scale=0, use_sigm
     return scale * q, scale, None
 
 
-def quant_weight_actor(weight, num_bits, scheme, v, min_scale, max_scale, use_sigmoid):
+def quant_weight_actor(weight, num_bits, scheme, v, min_scale, max_scale):
     """Quantizes and dequantizes weight symmetrically or asymmetrically .
 
     Args:
@@ -165,22 +116,12 @@ def quant_weight_actor(weight, num_bits, scheme, v, min_scale, max_scale, use_si
     """
     assert num_bits > 0, "num_bits should be larger than 0"
     if scheme == "sym":
-        return quant_weight_sym(weight, num_bits, v, min_scale, max_scale, use_sigmoid)
+        return quant_weight_sym(weight, num_bits, v, min_scale, max_scale)
     else:
-        return quant_weight_asym(weight, num_bits, v, min_scale, max_scale, use_sigmoid)
+        return quant_weight_asym(weight, num_bits, v, min_scale, max_scale)
 
 
-def quant_weight(
-    weight,
-    num_bits=4,
-    group_size=-1,
-    scheme="asym",
-    v=0,
-    min_scale=0,
-    max_scale=0,
-    use_sigmoid=False,
-    return_scalezp=False,
-):
+def quant_weight(weight, num_bits=4, group_size=-1, scheme="asym", v=0, min_scale=0, max_scale=0):
     """Quantizes and dequantizes weight, handing the group size issue .
 
     Args:
@@ -191,25 +132,12 @@ def quant_weight(
         v: Rounding value perturbation
         min_scale: Minimum scale coefficient for weight
         max_scale: Maximum scale coefficient for weight
-        use_sigmoid: Boolean indicating whether to use sigmoid; found useful in some scenarios for adam
 
     Returns:
         Quantized and dequantized weight, scale, zero-point
     """
     if group_size == -1 or weight.shape[1] < group_size:
-        weight, scale, zp = quant_weight_actor(
-            weight,
-            num_bits,
-            scheme=scheme,
-            v=v,
-            min_scale=min_scale,
-            max_scale=max_scale,
-            use_sigmoid=use_sigmoid,
-        )
-        if return_scalezp:
-            return weight, scale, zp
-
-        return weight
+        return quant_weight_actor(weight, num_bits, scheme=scheme, v=v, min_scale=min_scale, max_scale=max_scale)
     orig_shape = weight.shape
     if weight.shape[1] % group_size == 0:
         weight = weight.reshape(-1, group_size)
@@ -217,129 +145,33 @@ def quant_weight(
             v = v.reshape(-1, group_size)
 
         weight, scale, zp = quant_weight_actor(
-            weight,
-            num_bits,
-            scheme=scheme,
-            v=v,
-            min_scale=min_scale,
-            max_scale=max_scale,
-            use_sigmoid=use_sigmoid,
+            weight, num_bits, scheme=scheme, v=v, min_scale=min_scale, max_scale=max_scale
         )
-
         weight = weight.reshape(orig_shape)
-        if return_scalezp:
-            return weight, scale, zp
-        return weight
+        return weight, scale, zp
+
     else:
-        split_index = weight.shape[1] // group_size * group_size
-        weight1 = weight[:, :split_index]
-        weight1 = weight1.reshape(-1, group_size)
+        pad_len = (weight.shape[1] + group_size - 1) // group_size * group_size - weight.shape[1]
+        weight_new = torch.nn.functional.pad(weight, (0, pad_len))
+        v = torch.nn.functional.pad(v, (0, pad_len))
+        weight_new = weight_new.reshape(-1, group_size)
         if isinstance(v, torch.Tensor):
-            v1 = v[:, :split_index]
-            v1 = v1.reshape(-1, group_size)
-            v2 = v[:, split_index:]
-        else:
-            v1 = 0
-            v2 = 0
-        if isinstance(min_scale, torch.Tensor):
-            min_scale = min_scale.view(weight.shape[0], -1)
-            max_scale = max_scale.view(weight.shape[0], -1)
-            min_scale_1 = min_scale[:, : weight.shape[1] // group_size]
-            min_scale_2 = min_scale[:, weight.shape[1] // group_size :]
-            max_scale_1 = max_scale[:, : weight.shape[1] // group_size]
-            max_scale_2 = max_scale[:, weight.shape[1] // group_size :]
-            min_scale_1 = min_scale_1.reshape(-1)
-            min_scale_2 = min_scale_2.reshape(-1)
-            max_scale_1 = max_scale_1.reshape(-1)
-            max_scale_2 = max_scale_2.reshape(-1)
-        else:
-            min_scale_1 = min_scale
-            min_scale_2 = min_scale
-            max_scale_1 = max_scale
-            max_scale_2 = max_scale
-
-        (
-            weight1,
-            scale1,
-            zp1,
-        ) = quant_weight_actor(
-            weight1,
-            num_bits,
-            scheme=scheme,
-            v=v1,
-            min_scale=min_scale_1,
-            max_scale=max_scale_1,
-            use_sigmoid=use_sigmoid,
+            v = v.reshape(-1, group_size)
+        weight_new, scale, zp = quant_weight_actor(
+            weight_new, num_bits, scheme=scheme, v=v, min_scale=min_scale, max_scale=max_scale
         )
-        weight1 = weight1.reshape(orig_shape[0], split_index)
-        weight2 = weight[:, split_index:]
-        weight2, scale2, zp2 = quant_weight_actor(
-            weight2,
-            num_bits,
-            scheme=scheme,
-            v=v2,
-            min_scale=min_scale_2,
-            max_scale=max_scale_2,
-            use_sigmoid=use_sigmoid,
-        )
-        weight = torch.cat([weight1, weight2], dim=1)  ##TODO not supported output group
-        if return_scalezp:
-            scale = torch.cat([scale1, scale2], dim=0)
-            zp = torch.cat([zp1, zp2], dim=0)
-            return weight, scale, zp
-        return weight
+        weight_new = weight_new.reshape(orig_shape[0], -1)
+
+        weight_new = weight_new[:, :-pad_len]
+        return weight_new, scale, zp
 
 
-class WrapperMultiblock(torch.nn.Module):
-    """Wrapper multiple blocks to one module."""
-
-    def __init__(self, module_list):
-        """
-        Init function
-        Args:
-            module_list: a list of consecutive modules
-        """
-        super(WrapperMultiblock, self).__init__()
-        self.layers = torch.nn.ModuleList(module_list)
-
-    def forward(self, x, **kwargs):
-        """
-        forward function
-        Args:
-            x: Input
-            **kwargs: Other inputs
-
-        Returns:
-            output
-
-        """
-        hidden_states = x
-        for idx, decoder_layer in enumerate(self.layers):
-            layer_outputs = decoder_layer(hidden_states, **kwargs)
-            hidden_states = layer_outputs
-            if isinstance(hidden_states, tuple) or isinstance(hidden_states, list):
-                hidden_states = layer_outputs[0]
-        return hidden_states
-
-
-def get_block_names(model):
+def round_ste(x: torch.Tensor):
     """
-    Get the block names for transformers-like networks, this may have issues
-    Args:
-        model: The model
-
-    Returns:
-        all the block names
-
+    cp from https://github.com/OpenGVLab/OmniQuant/blob/main/quantize/quantizer.py
+    Implement Straight-Through Estimator for rounding operation.
     """
-    block_names = []
-    target_m = None
-    for n, m in model.named_modules():
-        if hasattr(type(m), "__name__") and "ModuleList" in type(m).__name__:
-            target_m = (n, m)
-    for n, m in target_m[1].named_children():
-        block_names.append(target_m[0] + "." + n)
-    return block_names
+    return (x.round() - x).detach() + x
 
 
 class SaveInputs:
@@ -402,8 +234,10 @@ class SaveInputs:
         for data in self.dataloader:
             if data is None:
                 continue
-
-            input_ids = data["input_ids"].to(self.model.device)
+            if isinstance(data, list):
+                input_ids = data.to(self.model.device)
+            else:
+                input_ids = data["input_ids"].to(self.model.device)
             if input_ids.shape[-1] < self.seqlen:
                 continue
             if total_cnt + input_ids.shape[0] > n_samples:
@@ -435,95 +269,54 @@ class SaveInputs:
                 break
 
 
-def round_ste(x: torch.Tensor):
+@torch.no_grad()
+def q_dq_weight(model: torch.nn.Module, num_bits=4, group_size=128, scheme="asym"):
+    target_m = None
+    for n, m in model.named_modules():
+        if hasattr(type(m), "__name__") and "ModuleList" in type(m).__name__:
+            target_m = (n, m)
+    block_names = []
+    for n, m in target_m[1].named_children():
+        block_names.append(target_m[0] + "." + n)
+    for name in block_names:
+        block = get_module(model, name)
+        for n, m in block.named_modules():
+            if isinstance(m, torch.nn.Linear):
+                m.weight.data.copy_(quant_weight(m.weight, num_bits=num_bits, group_size=group_size, scheme=scheme)[0])
+
+
+def get_module(model, key):
+    """Get module from model by key name.
+
+    Args:
+        model (torch.nn.Module): original model
+        key (str): module name to be replaced
     """
-    cp from https://github.com/OpenGVLab/OmniQuant/blob/main/quantize/quantizer.py
-    Implement Straight-Through Estimator for rounding operation.
+    module = model
+    name_list = key.split(".")
+    for name in name_list:
+        if hasattr(module, name):
+            module = getattr(module, name)
+            module = module
+    return module
+
+
+def set_module(model, key, new_module):
+    """Set new module into model by key name.
+
+    Args:
+        model (torch.nn.Module): original model
+        key (str): module name to be replaced
+        new_module (torch.nn.Module): new module to be inserted
     """
-    return (x.round() - x).detach() + x
-
-
-def sampling_inputs(input_ids, input_others, indices, seqlen):
-    if len(input_ids.shape) == 3:
-        if int(len(input_others["positional_inputs"]) > 0):
-            current_input_ids = input_ids[:, indices, :]
+    module = model
+    name_list = key.split(".")
+    for name in name_list[:-1]:
+        if hasattr(module, name):
+            module = getattr(module, name)
         else:
-            current_input_ids = input_ids[indices, :, :]
-    else:
-        n_samples = input_ids.shape[0] // seqlen
-        current_input_ids = input_ids.view(n_samples, seqlen, -1)
-        current_input_ids = current_input_ids[indices, :, :]
-        current_input_ids = current_input_ids.reshape(-1, input.shape[-1])
-
-    current_input_others = {}
-    current_input_others["positional_inputs"] = input_others["positional_inputs"]
-    for key in input_others.keys():
-        if "attention_mask" in key or "alibi" in key:
-            current_input_others[key] = None
-            if input_others[key] is not None:
-                current_input_others[key] = input_others[key][indices, ...]
-        else:
-            current_input_others[key] = input_others[key]
-
-    return current_input_ids, current_input_others
-
-
-# def move_to_device(input_ids, inputs_others=None, device=torch.device("cpu")):
-#     if input_ids != None:
-#         input_ids = input_ids.to(device)
-#     if inputs_others is not None:
-#         for key in inputs_others.keys():
-#             inputs_others[key] = inputs_others[key].to(device)
-#         return input_ids, inputs_others
-#     return input_ids
-
-
-def move_input_to_device(input, device=torch.device("cpu")):
-    if isinstance(input, torch.Tensor):
-        return input.to(device)
-    if isinstance(input, dict) or isinstance(input, UserDict):
-        for inp in input.keys():
-            input[inp] = move_input_to_device(input[inp], device)
-    elif isinstance(input, list) or isinstance(input, tuple):
-        input_res = []
-        for inp in input:
-            input_res.append(move_input_to_device(inp, device))
-        input = input_res
-    return input
-
-
-def block_forward(block, input_ids, input_others, amp=False, device=torch.device("cpu")):
-    if input_ids.device != device:
-        # input_ids, input_others = move_to_device(input_ids, input_others, device)
-        input_ids = move_input_to_device(input_ids, device)
-        input_others = move_input_to_device(input_others, device)
-    if "alibi" in input_others.keys():
-        attention_mask = input_others["attention_mask"]
-        alibi = input_others["alibi"]
-        alibi = alibi.reshape(-1, alibi.shape[2], alibi.shape[3])
-        if amp and device != torch.device("cpu"):
-            with autocast(device_type="cuda"):
-                output = block(
-                    input_ids, attention_mask=attention_mask, alibi=alibi
-                )  ##TODO is this correct for all models with alibi?
-        elif amp and device == torch.device("cpu"):
-            with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
-                output = block(input_ids, attention_mask=attention_mask, alibi=alibi)
-        else:
-            output = block(input_ids, attention_mask=attention_mask, alibi=alibi)
-    else:
-        input_tuple = input_others.pop("positional_inputs", None)
-        if amp and device != torch.device("cpu"):
-            with autocast(device_type="cuda"):
-                output = block.forward(input_ids, *input_tuple, **input_others)
-        elif amp and device == torch.device("cpu"):
-            with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
-                output = block.forward(input_ids, *input_tuple, **input_others)
-        else:
-            output = block.forward(input_ids, *input_tuple, **input_others)
-    if isinstance(output, list) or isinstance(output, tuple):
-        output = output[0]
-    return output
+            module = module
+    setattr(module, name_list[-1], new_module)
 
 
 def get_scale_shape(weight, group_size):
@@ -533,6 +326,60 @@ def get_scale_shape(weight, group_size):
         shape = weight.shape[0] * ((weight.shape[1] + group_size - 1) // group_size)
 
     return shape
+
+
+class WrapperLinear(torch.nn.Module):
+    def __init__(self, orig_layer, enable_minmax_tuning=True):
+        super(WrapperLinear, self).__init__()
+        self.orig_layer = orig_layer
+        self.num_bits = self.orig_layer.bits
+        self.group_size = self.orig_layer.group_size
+        self.scheme = (self.orig_layer.scheme,)
+        self.value = torch.nn.Parameter(
+            torch.zeros(self.orig_layer.weight.shape, device=self.orig_layer.weight.device), requires_grad=True
+        )
+        self.enable_minmax_tuning = enable_minmax_tuning
+        shape = get_scale_shape(self.orig_layer.weight, self.group_size)
+
+        if self.enable_minmax_tuning:
+            self.min_scale = torch.nn.Parameter(
+                torch.zeros(shape, device=self.orig_layer.weight.device), requires_grad=True
+            )
+            self.max_scale = torch.nn.Parameter(
+                torch.zeros(shape, device=self.orig_layer.weight.device), requires_grad=True
+            )
+        else:
+            self.min_scale = torch.tensor(0, device=self.orig_layer.weight.device)
+            self.max_scale = torch.tensor(0, device=self.orig_layer.weight.device)
+
+    def unwrapper(self, v, min_scale, max_scale):
+        min_scale.clamp_(-1, 0)
+        max_scale.clamp_(-1, 0)
+
+        q_dq_weight, scale, zp = quant_weight(
+            self.orig_layer.weight,
+            self.num_bits,
+            self.group_size,
+            self.scheme,
+            v,
+            min_scale,
+            max_scale,
+        )
+        self.orig_layer.weight.data.copy_(q_dq_weight)
+        self.orig_layer.weight.grad = None  ##clear grad
+        self.orig_layer.scale = scale.to("cpu")
+        self.orig_layer.zp = zp.to("cpu")
+        return self.orig_layer
+
+    def forward(self, x):
+        weight = self.orig_layer.weight
+        self.min_scale.data.copy_(torch.clamp(self.min_scale.data, -1, 0))
+        self.max_scale.data.copy_(torch.clamp(self.max_scale.data, -1, 0))
+        weight_q, _, _ = quant_weight(
+            weight, self.num_bits, self.group_size, self.scheme, self.value, self.min_scale, self.max_scale
+        )
+        weight_q = weight_q.to(weight.dtype)
+        return F.linear(x, weight_q, self.orig_layer.bias)
 
 
 class WrapperTransformerConv1d(torch.nn.Module):
@@ -555,19 +402,17 @@ class WrapperTransformerConv1d(torch.nn.Module):
             self.min_scale = torch.tensor(0, device=device)
             self.max_scale = torch.tensor(0, device=device)
 
-    def unwrapper(self, grad, min_scale_grad, max_scale_grad):
-        min_scale_grad.clamp_(-1, 0)
-        max_scale_grad.clamp_(-1, 0)
+    def unwrapper(self, v, min_scale, max_scale):
+        min_scale.clamp_(-1, 0)
+        max_scale.clamp_(-1, 0)
         weight_q, scale, zp = quant_weight(
             self.weight_t,
             self.num_bits,
             self.group_size,
             self.scheme,
-            grad,
-            min_scale_grad,
-            max_scale_grad,
-            use_sigmoid=self.use_sigmoid,
-            return_scalezp=True,
+            v,
+            min_scale,
+            max_scale,
         )
         self.orig_layer.weight.data.copy_(weight_q.t())
         self.orig_layer.weight.grad = None
@@ -579,7 +424,7 @@ class WrapperTransformerConv1d(torch.nn.Module):
         with torch.no_grad():
             self.min_scale.clamp_(-1, 0)
             self.max_scale.clamp_(-1, 0)
-        weight_q = quant_weight(
+        weight_q, _, _ = quant_weight(
             self.weight_t,
             self.num_bits,
             self.group_size,
@@ -587,153 +432,173 @@ class WrapperTransformerConv1d(torch.nn.Module):
             self.value,
             self.min_scale,
             self.max_scale,
-            use_sigmoid=self.use_sigmoid,
         )
+        weight_q = weight_q.to(self.weight_t.dtype)
         size_out = x.size()[:-1] + (self.orig_layer.nf,)
         x = torch.addmm(self.orig_layer.bias, x.view(-1, x.size(-1)), weight_q.t())
         x = x.view(*size_out)
         return x
 
 
-class WrapperLinear(torch.nn.Module):
-    def __init__(self, orig_layer, num_bits, group_size, scheme, use_sigmoid, enable_minmax_tuning=True):
-        super(WrapperLinear, self).__init__()
-        self.orig_layer = orig_layer
-        self.num_bits = num_bits
-        self.group_size = group_size
-        self.scheme = scheme
-        self.use_sigmoid = use_sigmoid
-        device = self.orig_layer.weight.device
-        self.value = torch.nn.Parameter(torch.zeros(self.orig_layer.weight.shape, device=device), requires_grad=True)
-        shape = get_scale_shape(self.orig_layer.weight, group_size)
-
-        if enable_minmax_tuning:
-            self.min_scale = torch.nn.Parameter(torch.zeros(shape, device=device), requires_grad=True)
-            self.max_scale = torch.nn.Parameter(torch.zeros(shape, device=device), requires_grad=True)
-        else:
-            self.min_scale = torch.tensor(0, device=device)
-            self.max_scale = torch.tensor(0, device=device)
-
-    def unwrapper(self, grad, min_scale_grad, max_scale_grad):
-        min_scale_grad.clamp_(-1, 0)
-        max_scale_grad.clamp_(-1, 0)
-
-        q_dq_weight, scale, zp = quant_weight(
-            self.orig_layer.weight,
-            self.num_bits,
-            self.group_size,
-            self.scheme,
-            grad,
-            min_scale_grad,
-            max_scale_grad,
-            use_sigmoid=self.use_sigmoid,
-            return_scalezp=True,
-        )
-        self.orig_layer.weight.data.copy_(q_dq_weight)
-        self.orig_layer.weight.grad = None  ##clear grad
-        self.orig_layer.scale = scale.to("cpu")
-        self.orig_layer.zp = zp.to("cpu")
-        return self.orig_layer
-
-    def forward(self, x):
-        weight = self.orig_layer.weight
-        with torch.no_grad():
-            self.min_scale.clamp_(-1, 0)
-            self.max_scale.clamp_(-1, 0)
-        weight_q = quant_weight(
-            weight,
-            self.num_bits,
-            self.group_size,
-            self.scheme,
-            self.value,
-            self.min_scale,
-            self.max_scale,
-            use_sigmoid=self.use_sigmoid,
-        )
-
-        return F.linear(x, weight_q, self.orig_layer.bias)
-
-
-def is_supported_layer(m, supported_types):
-    supported = False
-    for t in supported_types:
-        if isinstance(m, t):
-            supported = True
-    return supported
-
-
-def check_is_float(data_type, num_bits, group_size, scheme):
-    if ("fp" in data_type or "float" in data_type) and num_bits >= 16:
-        return True
-    return False
-
-
-def wrapper_block(block, enable_minmax_tuning, supported_types, use_sigmoid):
+def wrapper_block(block, enable_minmax_tuning):
     for n, m in block.named_modules():
-        if not is_supported_layer(m, supported_types):
-            continue
-        data_type = m.data_type
-        num_bits = m.bits
-        group_size = m.group_size
-        scheme = m.scheme
-        if check_is_float(data_type, num_bits, group_size, scheme):
-            continue
         if isinstance(m, torch.nn.Linear):
-            new_m = WrapperLinear(
-                m, num_bits, group_size, scheme, enable_minmax_tuning=enable_minmax_tuning, use_sigmoid=use_sigmoid
-            )
-            set_module(block, n, new_m)
-        elif isinstance(m, torch.nn.Conv1d):
-            pass
-        elif "Conv1D" in str(type(m)):
-            from transformers.modeling_utils import Conv1D
-
-            new_m = WrapperTransformerConv1d(
-                m, num_bits, group_size, scheme, enable_minmax_tuning=enable_minmax_tuning, use_sigmoid=use_sigmoid
-            )
+            new_m = WrapperLinear(m, enable_minmax_tuning=enable_minmax_tuning)
             set_module(block, n, new_m)
 
 
 @torch.no_grad()
-def unwrapper_block(block, grads, min_scale_grads, max_scale_grads):
+def unwrapper_block(block, vs, min_scales, max_scales):
     for n, m in block.named_modules():
         if isinstance(m, WrapperLinear) or isinstance(m, WrapperTransformerConv1d):
-            orig_layer = m.orig_layer
-            grad = 0
-            min_scale_grad = 0
-            max_scale_grad = 0
-            if isinstance(grads, dict):
-                grad = grads[n]
-            if isinstance(min_scale_grads, dict):
-                min_scale_grad = min_scale_grads[n]
-                min_scale_grad = torch.clamp(min_scale_grad, -1, 0)
-            if isinstance(max_scale_grads, dict):
-                max_scale_grad = max_scale_grads[n]
-                max_scale_grad = torch.clamp(max_scale_grad, -1, 0)
-            orig_layer = m.unwrapper(grad, min_scale_grad, max_scale_grad)
+            v = 0
+            min_scale = 0
+            max_scale = 0
+            if isinstance(vs, dict):
+                v = vs[n]
+            if isinstance(min_scales, dict):
+                min_scale = min_scales[n]
+                min_scale = torch.clamp(min_scale, -1, 0)
+            if isinstance(max_scales, dict):
+                max_scale = max_scales[n]
+                max_scale = torch.clamp(max_scale, -1, 0)
+            orig_layer = m.unwrapper(v, min_scale, max_scale)
             set_module(block, n, orig_layer)
 
 
-def collect_round_grad(block):
-    grads = {}
+def sampling_inputs(input_ids, input_others, indices, seqlen):
+    if len(input_ids.shape) == 3:
+        if int(len(input_others["positional_inputs"]) > 0):
+            current_input_ids = input_ids[:, indices, :]
+        else:
+            current_input_ids = input_ids[indices, :, :]
+    else:
+        n_samples = input_ids.shape[0] // seqlen
+        current_input_ids = input_ids.view(n_samples, seqlen, -1)
+        current_input_ids = current_input_ids[indices, :, :]
+        current_input_ids = current_input_ids.reshape(-1, input.shape[-1])
+
+    current_input_others = {"positional_inputs": input_others["positional_inputs"]}
+    for key in input_others.keys():
+        if "attention_mask" in key or "alibi" in key:
+            current_input_others[key] = None
+            if input_others[key] is not None:
+                current_input_others[key] = input_others[key][indices, ...]
+        else:
+            current_input_others[key] = input_others[key]
+
+    return current_input_ids, current_input_others
+
+
+def move_input_to_device(input, device=torch.device("cpu")):
+    if isinstance(input, torch.Tensor):
+        return input.to(device)
+    if isinstance(input, dict) or isinstance(input, UserDict):
+        for inp in input.keys():
+            input[inp] = move_input_to_device(input[inp], device)
+    elif isinstance(input, list) or isinstance(input, tuple):
+        input_res = []
+        for inp in input:
+            input_res.append(move_input_to_device(inp, device))
+        input = input_res
+    return input
+
+
+def block_forward(block, input_ids, input_others, amp=False, amp_dtype=torch.float16, device=torch.device("cpu")):
+    if input_ids.device != device:
+        # input_ids, input_others = move_to_device(input_ids, input_others, device)
+        input_ids = move_input_to_device(input_ids, device)
+        input_others = move_input_to_device(input_others, device)
+    if "alibi" in input_others.keys():
+        attention_mask = input_others["attention_mask"]
+        alibi = input_others["alibi"]
+        alibi = alibi.reshape(-1, alibi.shape[2], alibi.shape[3])
+        if amp and device != torch.device("cpu"):
+            with autocast(device_type="cuda", dtype=amp_dtype):
+                output = block(
+                    input_ids, attention_mask=attention_mask, alibi=alibi
+                )  ##TODO is this correct for all models with alibi?
+        elif amp and device == torch.device("cpu"):
+            with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
+                output = block(input_ids, attention_mask=attention_mask, alibi=alibi)
+        else:
+            output = block(input_ids, attention_mask=attention_mask, alibi=alibi)
+    else:
+        input_tuple = input_others.pop("positional_inputs", None)
+        if amp and device != torch.device("cpu"):
+            with autocast(device_type="cuda", dtype=amp_dtype):
+                output = block.forward(input_ids, *input_tuple, **input_others)
+        elif amp and device == torch.device("cpu"):
+            with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
+                output = block.forward(input_ids, *input_tuple, **input_others)
+        else:
+            output = block.forward(input_ids, *input_tuple, **input_others)
+    if isinstance(output, list) or isinstance(output, tuple):
+        output = output[0]
+    return output
+
+
+def collect_round_v(block):
+    vs = {}
     for n, m in block.named_modules():
-        if isinstance(m, WrapperLinear) or isinstance(m, WrapperTransformerConv1d):
-            grad = m.value.data
-            grads[n] = copy.deepcopy(grad)
-    return grads
+        if isinstance(m, WrapperLinear):
+            v = m.value.data
+            vs[n] = copy.deepcopy(v)
+    return vs
 
 
-def collect_minmax_grad(block):
-    min_grads = {}
-    max_grads = {}
+def collect_minmax_scale(block):
+    min_scales = {}
+    max_scales = {}
     for n, m in block.named_modules():
-        if isinstance(m, WrapperLinear) or isinstance(m, WrapperTransformerConv1d):
-            min_grads[n] = copy.deepcopy(torch.clamp(m.min_scale.data, -1, 0))
-            max_grads[n] = copy.deepcopy(torch.clamp(m.max_scale.data, -1, 0))
-    return min_grads, max_grads
+        if isinstance(m, WrapperLinear):
+            min_scales[n] = copy.deepcopy(torch.clamp(m.min_scale.data, -1, 0))
+            max_scales[n] = copy.deepcopy(torch.clamp(m.max_scale.data, -1, 0))
+    return min_scales, max_scales
 
 
-class OPTRoundQuantizer(object):
+@torch.no_grad()
+def get_batch_dim(input_others):
+    dim = int(len(input_others["positional_inputs"]) > 0)
+    return dim
+
+
+class WrapperMultiblock(torch.nn.Module):
+    def __init__(self, module_list):
+        super(WrapperMultiblock, self).__init__()
+        self.layers = torch.nn.ModuleList(module_list)
+
+    def forward(self, x, **kwargs):
+        hidden_states = x
+        for idx, decoder_layer in enumerate(self.layers):
+            layer_outputs = decoder_layer(hidden_states, **kwargs)
+            hidden_states = layer_outputs
+            if isinstance(hidden_states, tuple) or isinstance(hidden_states, list):
+                hidden_states = layer_outputs[0]
+        return hidden_states
+
+
+def get_block_names(model):
+    """
+    Get the block names for transformers-like networks, this may have issues
+    Args:
+        model: The model
+
+    Returns:
+        all the block names
+
+    """
+    block_names = []
+    target_m = None
+    for n, m in model.named_modules():
+        if hasattr(type(m), "__name__") and "ModuleList" in type(m).__name__:
+            target_m = (n, m)
+    for n, m in target_m[1].named_children():
+        block_names.append(target_m[0] + "." + n)
+    return block_names
+
+
+class AutoRound(object):
     def __init__(
         self,
         model,
@@ -746,7 +611,6 @@ class OPTRoundQuantizer(object):
         bs: int = 8,
         amp: bool = True,
         device="cuda:0",
-        optimizer=None,
         lr_scheduler=None,
         dataloader=None,  ## to support later
         default_dataset_name: str = "NeelNanda/pile-10k",
@@ -766,7 +630,6 @@ class OPTRoundQuantizer(object):
         not_use_mse: bool = False,
         dynamic_max_gap: int = -1,
         data_type: str = "int",  ##only support data_type
-        use_sigmoid=None,
         **kwargs,
     ):
         """
@@ -829,6 +692,7 @@ class OPTRoundQuantizer(object):
         if self.device == "cpu":
             self.amp_dtype = torch.bfloat16
         if self.amp:
+            self.model = self.model.to(self.amp_dtype)
             logger.info(f"using {self.amp_dtype}")
 
         if dataloader is None:
@@ -844,31 +708,33 @@ class OPTRoundQuantizer(object):
         self.dynamic_max_gap = dynamic_max_gap
         self.enable_full_range = enable_full_range
         assert self.enable_full_range is False, "only support enable_full_range=False currently"
-        self.optimizer = optimizer
-        self.scale_grad = True
-        self.use_sigmoid = use_sigmoid
-
-        if self.optimizer is None:
-            self.use_sigmoid = False
-            self.scale_grad = False
-            from .sign_sgd import SGD
-
-            self.optimizer = SGD
-
-        elif isinstance(self.optimizer, str):
-            self.optimizer = getattr(torch.optim, self.optimizer)
-        else:
-            self.optimizer = optimizer
-        if self.use_sigmoid is None:
-            self.use_sigmoid = False
         self.lr_scheduler = lr_scheduler
         self.set_layerwise_config(self.weight_config)
+        self.optimizer = self.get_optimizer(None)
+
+    def get_optimizer(self, optimizer):
+        from sign_sgd import SGD
+
+        return SGD
+
+    def get_scaler(self):
+        return None
+
+    def scale_loss_and_backward(self, scaler, loss):
+        scale_loss = loss * 1000
+        scale_loss.backward()
+        return scale_loss
+
+    def step(self, scaler, optimizer, lr_schedule):
+        optimizer.step()
+        optimizer.zero_grad()
+        lr_schedule.step()
 
     def check_configs(self):
         assert isinstance(self.model, torch.nn.Module)
         assert self.bits > 0, "bits must be positive"
         assert self.group_size == -1 or self.group_size >= 1, "only supports positive group_size or -1(per channel)"
-        assert self.bs > 0, "batch size must be positive"
+        assert self.train_bs > 0, "batch size must be positive"
         assert self.iters > 0, "iters must be positive"
         assert self.seqlen > 0, "seqlen must be positive"
         assert self.n_blocks > 0, "n_blocks must be positive"
@@ -935,7 +801,38 @@ class OPTRoundQuantizer(object):
         calib_dataset = calib_dataset.map(self.default_tokenize_function, batched=True)
         calib_dataset.set_format(type="torch", columns=["input_ids"])
         calib_dataloader = DataLoader(calib_dataset, batch_size=self.train_bs, shuffle=False, collate_fn=collate_batch)
+
         return calib_dataloader
+
+    # def get_default_dataloader(self, data_name="NeelNanda/pile-10k"):
+    #     from datasets import load_dataset
+    #
+    #     seqlen = self.seqlen
+    #     calib_dataset = load_dataset(data_name, split=self.dataset_split).shuffle(seed=self.seed)
+    #     samples = []
+    #     cnt = 0
+    #     tmp_samples = []
+    #     for data in calib_dataset:
+    #         line = data["text"]
+    #         line = line.strip()
+    #         line_tokenized = self.tokenizer.encode(line)
+    #         if len(line_tokenized) < self.seqlen:
+    #             continue
+    #         index = 0
+    #         sample = line_tokenized[index:index + self.seqlen]
+    #         if sample.count(sample[-1]) > seqlen // 2:
+    #             continue
+    #         sample = torch.tensor(sample)
+    #         cnt += 1
+    #         tmp_samples.append(sample)
+    #         if cnt % self.train_bs == 0:
+    #             tmp = torch.vstack(tmp_samples)
+    #             samples.append(tmp)
+    #             tmp_samples = []
+    #         if cnt >= self.n_samples:
+    #             break
+    #
+    #     return samples
 
     def get_batch_dim(self, input_others):
         dim = int(len(input_others["positional_inputs"]) > 0)
@@ -947,44 +844,41 @@ class OPTRoundQuantizer(object):
         for i in range(0, self.n_samples, bs):
             indices = torch.arange(i, i + bs).to(torch.long)
             tmp_input_ids, tmp_input_others = sampling_inputs(input_ids, input_others, indices, self.seqlen)
-            tmp_output = block_forward(block, tmp_input_ids, tmp_input_others, self.amp, device).to(cache_device)
+            tmp_output = block_forward(block, tmp_input_ids, tmp_input_others, self.amp, self.amp_dtype, device).to(
+                cache_device
+            )
             output.append(tmp_output)
         output = torch.cat(output, dim=batch_dim)
         torch.cuda.empty_cache()
         return output
 
-    def loss_scale_and_backward(self, scaler, loss):
-        if scaler is not None:
-            scale_loss = scaler.scale(loss)
-            scale_loss.backward()
-            return scale_loss
-        else:
-            loss *= 1000
-            loss.backward()
-            return loss
-
-    def step(self, scaler, optimizer, lr_schedule):  ##TODO signround does not need this
-        if scaler is not None:
-            scaler.step(optimizer)
-            optimizer.zero_grad()
-            lr_schedule.step()
-            scaler.update()
-        else:
-            optimizer.step()
-            optimizer.zero_grad()
-            lr_schedule.step()
-
-    def is_supported_type(self, m):
-        return hasattr(m, "orig_layer")
+    # def loss_scale_and_backward(self, scaler, loss):
+    #     if scaler is not None:
+    #         scale_loss = scaler.scale(loss)
+    #         scale_loss.backward()
+    #         return scale_loss
+    #     else:
+    #         loss *= 1000
+    #         loss.backward()
+    #         return loss
+    #
+    # def step(self, scaler, optimizer, lr_schedule):  ##TODO signround does not need this
+    #     if scaler is not None:
+    #         scaler.step(optimizer)
+    #         optimizer.zero_grad()
+    #         lr_schedule.step()
+    #         scaler.update()
+    #     else:
+    #         optimizer.step()
+    #         optimizer.zero_grad()
+    #         lr_schedule.step()
+    #
 
     def quant_block(self, block, input_ids, input_others, q_input=None, device=torch.device("cpu")):
-        batch_dim = self.get_batch_dim(input_others)
+        batch_dim = get_batch_dim(input_others)
         if not self.low_gpu_mem_usage and input_ids.device != device:
-            # input_ids, input_others = move_to_device(input_ids, input_others, device)
-            input_ids = move_input_to_device(input_ids, device)  ##move input data to GPU
+            input_ids = move_input_to_device(input_ids, device)
             input_others = move_input_to_device(input_others, device)
-        torch.cuda.empty_cache()
-
         cache_device = device
         if self.low_gpu_mem_usage:
             cache_device = "cpu"
@@ -993,17 +887,16 @@ class OPTRoundQuantizer(object):
         if q_input is not None:
             input_ids = q_input.to(cache_device)
 
-        wrapper_block(block, self.enable_minmax_tuning, self.supported_types, use_sigmoid=self.use_sigmoid)
+        wrapper_block(block, self.enable_minmax_tuning)
 
-        best_loss = torch.finfo(torch.float).max
-        mse_loss = torch.nn.MSELoss()
         round_params = []
         minmax_params = []
         for n, m in block.named_modules():
-            if self.is_supported_type(m):
+            if hasattr(m, "orig_layer"):
                 round_params.append(m.value)
                 minmax_params.append(m.min_scale)
                 minmax_params.append(m.max_scale)
+
         if self.enable_minmax_tuning:
             optimizer = self.optimizer(
                 [{"params": round_params}, {"params": minmax_params, "lr": self.minmax_lr}], lr=self.lr, weight_decay=0
@@ -1017,22 +910,21 @@ class OPTRoundQuantizer(object):
             )
         else:
             lr_schedule = copy.deepcopy(self.lr_scheduler)
+
+        pick_samples = self.train_bs
         if len(input_ids.shape) == 3:
             n_samples = input_ids.shape[0]
         else:
             n_samples = input_ids.shape[0] // self.seqlen
         if self.sampler != "rand":
-            indices = torch.randperm(n_samples)[: self.train_bs]
+            indices = torch.randperm(n_samples)[:pick_samples]
         last_best_iter = 0
-        scaler = None
-        if self.amp and self.scale_grad:
-            from torch.cuda.amp import GradScaler
-
-            scaler = GradScaler(init_scale=1024, growth_interval=100000)
-
+        best_loss = torch.finfo(torch.float).max
+        mse_loss = torch.nn.MSELoss().to(device)
+        scaler = self.get_scaler()
         for i in range(self.iters):
             if self.sampler == "rand":
-                indices = torch.randperm(n_samples)[: self.train_bs]
+                indices = torch.randperm(n_samples)[:pick_samples]
 
             total_loss = 0
             for _ in range(self.gradient_accumulate_steps):
@@ -1052,36 +944,40 @@ class OPTRoundQuantizer(object):
                     current_output = current_output.reshape(-1, current_output.shape[-1])
                 current_output = move_input_to_device(current_output, device)
 
-                output_q = block_forward(block, current_input_ids, current_input_others, self.amp, device)
+                output_q = block_forward(
+                    block, current_input_ids, current_input_others, self.amp, self.amp_dtype, device
+                )
                 if self.amp and device != torch.device("cpu"):
-                    with autocast(device_type="cuda"):
+                    with autocast(device_type="cuda", dtype=self.amp_dtype):
                         loss = mse_loss(output_q, current_output)
                 elif self.amp:
                     with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
                         loss = mse_loss(output_q, current_output)
                 else:
                     loss = mse_loss(output_q, current_output)
-                loss = loss / self.gradient_accumulate_steps
-                total_loss += loss.item()
-                scale_loss = self.loss_scale_and_backward(scaler, loss)
+
+                total_loss += (
+                    loss.item() / self.gradient_accumulate_steps
+                )  ##TODO gradient accumulate step for other optimizer
+                self.scale_loss_and_backward(scaler, loss)
+                # loss.backward()
 
             if total_loss < best_loss:
                 best_loss = total_loss
                 if not self.not_use_mse:
                     # print(f"get better result at iter {i}, the loss is {total_loss}", flush=True)
-                    best_grad = collect_round_grad(block)
-                    best_min_scale_grad, best_max_scale_grad = collect_minmax_grad(block)
+                    best_v = collect_round_v(block)
+                    best_min_scale, best_max_scale = collect_minmax_scale(block)
                     last_best_iter = i
             if self.not_use_mse and i == self.iters - 1:
-                best_grad = collect_round_grad(block)
-                best_min_scale_grad, best_max_scale_grad = collect_minmax_grad(block)
+                best_v = collect_round_v(block)
+                best_min_scale, best_max_scale = collect_minmax_scale(block)
 
             if not self.not_use_mse:
                 if self.dynamic_max_gap > 0 and i - last_best_iter >= self.dynamic_max_gap:
                     break
             self.step(scaler, optimizer, lr_schedule)
-
-        unwrapper_block(block, best_grad, best_min_scale_grad, best_max_scale_grad)
+        unwrapper_block(block, best_v, best_min_scale, best_max_scale)
         if self.use_quant_input:
             q_outputs = self.get_block_outputs(
                 block, input_ids, input_others, self.train_bs, device, cache_device, batch_dim
@@ -1147,7 +1043,7 @@ class OPTRoundQuantizer(object):
 
         if self.amp:
             self.model = self.model.to(self.amp_dtype)
-        if not self.low_gpu_mem_usage:  ##TODO automaticall detect
+        if not self.low_gpu_mem_usage:  ##TODO automatically detect
             self.model = self.model.to(self.device)
 
         save_input_actor = SaveInputs(self.model, self.dataloader, self.seqlen, block_names[0])
@@ -1170,3 +1066,104 @@ class OPTRoundQuantizer(object):
                     delattr(m, "scale")
                     delattr(m, "zp")
         return self.model, self.weight_config
+
+
+class AutoOPTRound(AutoRound):
+    def __init__(
+        self,
+        model,
+        tokenizer=None,
+        optimizer="AdamW",
+        bits: int = 4,
+        group_size: int = 128,
+        scheme: str = "asym",
+        weight_config: dict = {},
+        enable_full_range: bool = False,
+        bs: int = 8,
+        amp: bool = True,
+        device="cuda:0",
+        lr_scheduler=None,
+        dataloader=None,
+        default_dataset_name: str = "NeelNanda/pile-10k",
+        dataset_split: str = "train",
+        use_quant_input: bool = True,
+        enable_minmax_tuning: bool = True,
+        lr: float = 0.005,
+        minmax_lr: float = 0.005,
+        low_gpu_mem_usage: bool = True,
+        iters: int = 200,
+        seqlen: int = 2048,
+        n_samples: int = 512,
+        sampler: str = "rand",
+        seed: int = 42,
+        n_blocks: int = 1,
+        gradient_accumulate_steps: int = 1,
+        not_use_mse: bool = False,
+        dynamic_max_gap: int = -1,
+        data_type: str = "int",
+        **kwargs,
+    ):
+        super(AutoOPTRound, self).__init__(
+            model,
+            tokenizer,
+            bits,
+            group_size,
+            scheme,
+            weight_config,
+            enable_full_range,
+            bs,
+            amp,
+            device,
+            lr_scheduler,
+            dataloader,
+            default_dataset_name,
+            dataset_split,
+            use_quant_input,
+            enable_minmax_tuning,
+            lr,
+            minmax_lr,
+            low_gpu_mem_usage,
+            iters,
+            seqlen,
+            n_samples,
+            sampler,
+            seed,
+            n_blocks,
+            gradient_accumulate_steps,
+            not_use_mse,
+            dynamic_max_gap,
+            data_type,
+            **kwargs,
+        )
+
+        self.optimizer = self.get_optimizer(optimizer)
+
+    def get_optimizer(self, optimizer):
+        if optimizer is None:
+            optimizer = torch.optim.AdamW
+            ##TODO need to force set lr to some better candidates
+
+        elif isinstance(optimizer, str):
+            optimizer = getattr(torch.optim, self.optimizer)
+        else:
+            optimizer = optimizer
+        return optimizer
+
+    def get_scaler(self):
+        scaler = None
+        if self.amp and self.scale_grad:
+            from torch.cuda.amp import GradScaler
+
+            scaler = GradScaler(init_scale=1024, growth_interval=100000)
+        return scaler
+
+    def scale_loss_and_backward(self, scaler, loss):
+        scale_loss = scaler.scale(loss)
+        scale_loss.backward()
+        return scale_loss
+
+    def step(self, scaler, optimizer, lr_schedule):
+        scaler.step(optimizer)
+        optimizer.zero_grad()
+        lr_schedule.step()
+        scaler.update()
